Refactor away a barrel file
