+++ app/features/tournament/core/sets.server.ts
@@ -1,7 +1,7 @@
 import type { Tables } from "~/db/tables";
 import type { ModeShort, StageId } from "~/modules/in-game-lists/types";
-import type { TournamentMaplistSource } from "~/modules/tournament-map-list-generator";
-import { sourceTypes } from "~/modules/tournament-map-list-generator";
+import { sourceTypes } from "~/modules/tournament-map-list-generator/constants";
+import type { TournamentMaplistSource } from "~/modules/tournament-map-list-generator/types";
 import invariant from "~/utils/invariant";
 import { logger } from "~/utils/logger";
 import { findRoundsByTournamentId } from "../queries/findRoundsByTournamentId.server";
