--- app/modules/tournament-map-list-generator/index.ts
@@ -1,9 +1,0 @@
-export { sourceTypes } from "./constants";
-export { createTournamentMapList } from "./tournament-map-list";
-export type {
-	BracketType,
-	TournamentMapListMap,
-	TournamentMaplistInput,
-	TournamentMaplistSource,
-} from "./types";
-export { seededRandom } from "./utils";
